Fixed some bugs in Upload Video and Delete Video and some client pages

+++ server/controller/channelController.js
@@ -15,9 +15,12 @@ exports.getChannel = async (req, res, next) => {
     }
 }
 
+// Create a new Channel
 exports.createChannel = async (req, res, next) => {
     try{
-        const channel = await Channel.create(req.body);
+        const { name, description } = req.body;
+        // const channel = await Channel.create(req.body);
+        const channel = await Channel.create({ name, description, user: req.user._id});
         // console.log(req.cookie.userid);
         res.status(201).json({
             status: 'success',
@@ -31,6 +34,7 @@ exports.createChannel = async (req, res, next) => {
     }
 }
 
+// Get all Channels of user
 exports.getAllChannel = async (req, res, next) => {
     try{
         // const channels = await Channel.find({ user: req.params.id });
@@ -50,12 +54,21 @@ exports.getAllChannel = async (req, res, next) => {
 
 exports.deleteChannel = async (req, res, next) => {
     try{
-        const channel = await Channel.findByIdAndDelete(req.params.id);
-        console.log(channel);
-        res.status(201).json({
-            status: 'success',
-            data: channel 
-        })
+        const isAdmin = await Channel.findOne({ _id: req.params.id, user: req.user._id });
+        console.log(isAdmin);
+        if(isAdmin){
+            const channel = await Channel.findByIdAndDelete(req.params.id);
+            console.log(channel);
+            res.status(201).json({
+                status: 'success',
+                data: channel 
+            })
+        } else{
+            res.status(401).json({
+                status: "failed",
+                data: "Your are not authorized to delete this channel"
+            });
+        }
     } catch(err) {
         res.status(500).json({
             status: 'failed',
